Update calculator using match-case statement.py

+++ Calculator/calculator using match-case statement.py
@@ -1,7 +1,7 @@
 # Simple calculator using match-case statement
 
-a=int(input("Enter first number: "))            # Get first number from user using the input function
-b=int(input("Enter second number: "))           # Get second number from user using the input function
+a=float(input("Enter first number: "))            # Get first number from user using the input function
+b=float(input("Enter second number: "))           # Get second number from user using the input function
 
 sym= input("What operation do you want to perform (+, -, *, /): ")       # Get the operation symbol from user  
 
@@ -16,5 +16,6 @@ match sym:                                  # Use match-case to handle different
         print(f"The quotient is: {a / b}")
     case _:
         print("Invalid operation.")
+
